app/requirements.txt
fastapi
uvicorn
mysql-connector-python
bcrypt
jinja2
python-multipart
